// src/pages/types.ts
export interface PermitInterface {
  id: string;
  permit_: string;
  permit_type: string;
  review_type: string;
  application_start_date: Date;
  issue_date: Date;
  processing_time: number;
  street_number: number;
  street_direction: string;
  street_name: string;
  suffix: string;
  work_description: string;
  building_fee_paid: number;
  zoning_fee_paid: number;
  other_fee_paid: number;
  subtotal_paid: number;
  building_fee_unpaid: number;
  zoning_fee_unpaid: number;
  other_fee_unpaid: number;
  subtotal_unpaid: number;
  building_fee_waived: number;
  zoning_fee_waived: number;
  other_fee_waived: number;
  subtotal_waived: number;
  total_fee: number;
  contact_1_type: string;
  contact_1_name: string;
  contact_1_city: string;
  contact_1_state: string;
  contact_1_zipcode: string;
  contact_2_type: string;
  contact_2_name: string;
  contact_2_city: string;
  contact_2_state: string;
  contact_2_zipcode: string;
  reported_cost: number;
  pin1: string;
  community_area: number;
  census_tract: number;
  ward: number;
  xcoordinate: number;
  ycoordinate: number;
}